Rename examples
